Floating Test/Semantic-Parsing-of-picture-descriptions/requirements.txt
Pillow
PySimpleGUI
numpy